refactor: reorganizar estructura del modulo payment
