+++ src/main/java/com/coworking/payment/config/StripeConfig.java
@@ -11,7 +11,7 @@ import org.springframework.context.annotation.Configuration;
 @RequiredArgsConstructor
 public class StripeConfig {
 
-    @Value("${stripe.secret-key}")
+    @Value("${webhook.secret-key}")
     private String secretKey;
 
     @PostConstruct
