+++ src/test/java/com/coworking/resources/service/reservation/ReservationServiceTest.java
@@ -242,7 +242,7 @@ class ReservationServiceTest {
     }
 
 
-    //stripe paid
+    //webhook paid
 
     @Test
     void shouldMarkReservationAsPaid() {
